Move functions to util
